refactor: Cleanup Input component.
