--- src/components/Guilds/common/Form/Input/index.ts
@@ -1,1 +1,0 @@
-export * from './Input';
