+++ src/components/Guilds/common/Form/NumericalInput.tsx
@@ -1,7 +1,7 @@
 // Based on https://github.com/levelkdev/dxswap-dapp/blob/master/src/components/Input/NumericalInput/index.tsx
 
 import React from 'react';
-import { Input } from '.';
+import Input from './Input';
 import { escapeRegExp } from '../../../../utils';
 
 const inputRegex = RegExp(`^\\d*(?:\\\\[.])?\\d*$`); // match escaped "." characters via in a non-capturing group
