--- src/components/Guilds/common/Form/index.ts
@@ -1,1 +1,0 @@
-export * from './Input';
