+++ src/pages/Guilds/LandingPage.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import styled from 'styled-components';
 import { Button } from 'components/Guilds/common/Button';
-import { Input } from 'components/Guilds/common/Form';
+import Input from 'components/Guilds/common/Form/Input';
 import { AiOutlineSearch } from 'react-icons/ai';
 import { Flex, Box } from 'components/Guilds/common/Layout';
 import { MdOutlinePeopleAlt } from 'react-icons/md';
